fix: Improve autosave and unsaved changes detection (#24102)
* docs: Remove outdated comment

We already sync the title value.

* fix: Improve autosave and unsaved changes detection

Detecting content changes by comparing the content received from
GutenbergKit to the latest server-provided post content is problematic
as the server makes slight modifications to content markup, creating
perpetually changed content. We now rely upon GutenbergKit's `changed`
parameter as the source of truth.

* fix: Increase autosave frequency and robustness

Embrace the existing `Autosaver` utility so the frequency is higher, as
the utility includes both a time interval and a number of changes count
which determine whether an autosave should occur. This also further
aligns GutenbergKit's behavior with Gutenberg Mobile.

* build: Update GutenbergKit ref

* build: Update GutenbergKit ref

* docs: Add release note

* build: Update GutenbergKit ref