Replace all `ContextManager` `sharedInstance()` usages with `shared` (#24187)
Apart from being more Swift-y, this will also allow us to differentiate
between a shared instance for the Objective-C layer (`CoreDataStack`)
and one for the Swift layer (`CoreDataStackSwift`). This differentiation
will be valuable if/when we'll replace `ContextManager` references with
`CoreDataStack`/`CoreDataStackSwift` so that WordPressDataObjC can
refer to the functionality without knowing about `ContextManager`, which
is defined in the Swift WordPressData layer.