Extract SFHFKeychainUtils to dedicated Swift package (#24192)
* Add `import UIKit` to some files shared with notification extension

All those Swift files compile without importing UIKit because
`SFHFKeychainUtils.h` did that for them and they got access to it via
the bridging header.

Next, we'll move SFHFKeychainUtils into the modules, and UIKit will no
longer be available through it.

Side note: UIKit should not have been available via SFHFKeychainUtils in
the first place. On top of that, SFHFKeychainUtils does not need UIKit
either.

* Extract SFHFKeychainUtils in a dedicated Swift package

We need it in a dedicated package because the Objective-C implementation
does not use ARC. Therefore, we need to use the `-fno-objc-arc` compiler
setting. But such a setting should be applied sparingly so instead of
blanket using it in an existing Objective-C package like
WordPressSharedObjC, we use it surgically just for this file.