Clobber all migrations mappings and policies — Very old (#24208)
* Clobber all migrations mappings and policies — Very old

For the sake of convenience and code maintenance, we decided to remove
these migrations because we consider it safe to assume all existing
users are on later versions. If no user will ever run the migrations,
the code is essentially dead code that we can remove.

For reference, the current version is 154 and the oldest migration was
91-to-92. The model version 92 dates back to 2019, see
dcdba23bf7a416b8c4a5a190e442cc6b2e856293

* Move a few more utils types to WordPressShared (#24210)

- KeychainUtils & co
- UserPersistentRepository & co
- LocalFileStore & co