Decouple some model implementations from the app targets - Round 2 (#24349)
* Move `supportsSiteManagementServices()` to `Blog+Capabilities`

* Provide a public `init()` to `BlogQuery`

* Remove `isEqual(to:)` usage in favor of `==` in `Comment`

I don't remember why in the context of the #24165 work this gave
problem, other than it wouldn't compile.

Given the types being compared as `String`, using Swift's `==` seems
like a better solution anyway.

* Replace `Bundle.main` in `ContextManager`

This unblocks finding the `xcdatamodeld` once both `ContextManager` and
the Core Data schema are part of the same framework.

* Implement an explicit `init` for `GravatarProfile`

This way, once it's in the WordPressData framework, it'll be accessible
by consumers.

* Split `Notification` "pure" model code from apps details

The `Notification` model is littered with functionality that has to do
with the view layer. To move `Notification` to WordPressData in a timely
fashion, we need to separate the model description from all that other
logic.

* Extract `NotificationSyncMediatorProtocol` to dedicated file

* Split `RevisionDiff` model definition from `NSAttributedString` logic

* Move `RevisionDiff` extension to `DiffAbstractValue+Attributes.swift`

That's where they're used, and they're not needed in WordPressData.

* Move `SiteOrganizationType` definition to dedicated type

It will go into WordPressData next

* Provide `UserProfile` with a public init

* Extract `authKeychainServiceName` in `WPAccount` to dedicated file

Again, this simplifies moving to WordPressData without having to carry
along additional dependencies