Replace `asURL()` call with `URL(string:)` (#24350)
* Replace `asURL()` call with `URL(string:)` in `WordPressOrgRestApi`

The method comes from an old Alamofire version. Since its
implementation, `URL(string:)` stopped failing on "invalid" strings such
as "invalid-url". Between the different behavior and the cost of an
additional dependency, the value of using that method is marginal at
best.

* Replace `asURL()` call with `URL(string:)` in `GutenbergMediaEditorImage`

The method comes from an old Alamofire version. Since its
implementation, `URL(string:)` stopped failing on "invalid" strings such
as "invalid-url". Between the different behavior and the cost of an
additional dependency, the value of using that method is marginal at
best.

* Directly decode `URL` without using `String` + `asURL()` combo

`asURL()` comes from an old Alamofire version. Since its
implementation, `URL(string:)` stopped failing on "invalid" strings such
as "invalid-url". Between the different behavior and the cost of an
additional dependency, the value of using that method is marginal at
best.

But aside from that, Swift now offers direct `URL` decoding, so we don't
need to jump through any hoop.

And notice that `StockPhotosMedia` has tests that validate the JSON
decoding.