Tweak a few file to allow WordPressData to compile (#24375)
* Move code `localIconImage` to app targets / Keystone

This way, when `PublicizeService` moves to WordPressData, it won't need
the dependency used in that code.

* Use `BuildSettings` over `AppConfiguration` in `SharedDataIssueSolver`

* Use `BuildSettings` over `AppConfiguration` in `WPAccount+Lookup`

* Use `_private_wordPressComRestApi` in `Blog.m` to work around deps

* Properly call `alloc` in `PostService.m`

* Move `Blog` `wordPressComRestApi` and consumers from ObjC to Swift

Apologies for the annoying diff due to the change from methods to
computed vars.

I suppose I could have left the implementations as method, but it felt
cleaner to move them to more Swift-idiomatic computed vars.