Add empty WordPressData target + custom test plan (#24374)
* Add empty WordPressData target + custom test plan

* Do not use a host app to run WordPressData tests

* Move WordPressData sources to Sources/WordPressData

* Move WordPressDataTests to Tests folder

* Remove WordPressData the Swift package as an Xcode target dependency

* Add dummy type to verify setup

* Use the same settings as Keystone for WordPressData

* fix tests

* Use the same build settings as Authenticator in WordPressDataTests

* Remove WordPressData from WordPress' dependencies

It's not time for that yet

* Adjust linking so that WordPress, Jetpack, and the tests build

The issue is that I named the framework and the package both
WordPressData and that we already have existing `import`s of it in
production and test code.

To keep the diff simple, I decided to leave those imports in place and
simply:

- Remove the WordPressData package (empty) dependency
- Add explicit WordPressDataObjc package (empty) dependency
- Add WordPressData framework (empty) dependency

This sorts the compilation and running locally.

As we'll move files to WordPressData and/or away from Keystone, we might
need to change this approach.

* Explicitly set code signing for WordPressData

See how this build
https://buildkite.com/automattic/wordpress-ios/builds/26971 failed

* Set provisioning style as manual in WordPressData