Remove WordPressDataObjC and WordPressData modules (#24402)
* Remove `WordPressDataObjC` module

It was good to have it there as a placeholder, but it's giving me some
odd compilation issue where the extensions seem to think they need it as
a dependency.

It was empty anyway, so effectively dead weight. It'll be easy enough to
recreate it when the time comes.

* Remove `WordPressData` module (the Swift package)

Same rationale as the `WordPressDataObjC` removal in the previous
commit. It was dead weight.

Plus, it could be confusing to have a module (albeit empty and unused)
and a framework with the same name.